chore(integrations): trim github callback pipeline diff
Reduce review surface on top of the existing github callback work.
No behaviour change.

- Drop link_existing and oauth_authorize endpoints; verified no callers
  in this repo or PostHog Code app.
- Drop legacy state-cache dual-write/read (github_state:{user},
  github_user_install_state:{token}); state TTL is short, no upgrade
  window needed.
- Merge auth.py + oauth.py + parse.py into router.py.
- Merge team_finish.py into team_services.py.
- Merge validation.py into types.py.
- Parameterize duplicated tests and hoist fixture helpers.
- Strip module docstrings that restate filenames and
  from __future__ import annotations (project is Python 3.12+).
- Drop dead invalidate_repository_cache method on GitHubIntegrationBase.
