+++ posthog/api/github_callback/state.py
@@ -1,7 +1,3 @@
-"""Server-side authorize state for GitHub setup callbacks."""
-
-from __future__ import annotations
-
 import os
 from typing import Any, cast
 from urllib.parse import parse_qsl, urlparse
@@ -14,8 +10,6 @@ from rest_framework.request import Request
 
 from posthog.api.github_callback.types import (
     GITHUB_AUTHORIZE_STATE_CACHE_TTL_SECONDS,
-    GITHUB_INSTALL_STATE_CACHE_PREFIX,
-    GITHUB_INSTALL_STATE_TTL_SECONDS,
     GITHUB_UNIFIED_AUTHORIZE_CACHE_PREFIX,
     GITHUB_UNIFIED_AUTHORIZE_PENDING_PREFIX,
     PERSONAL_INTEGRATIONS_SETTINGS_PATH,
@@ -35,20 +29,12 @@ def parse_github_authorize_state_param(state_raw: str | None) -> tuple[str | Non
     return state_raw, None
 
 
-def github_authorize_state_cache_key(user_id: int) -> str:
-    return f"github_state:{user_id}"
-
-
 def unified_authorize_cache_key(token: str) -> str:
     return f"{GITHUB_UNIFIED_AUTHORIZE_CACHE_PREFIX}{token}"
 
 
 def unified_authorize_pending_cache_key(user_id: int) -> str:
     return f"{GITHUB_UNIFIED_AUTHORIZE_PENDING_PREFIX}{user_id}"
-
-
-def legacy_install_state_cache_key(token: str) -> str:
-    return f"{GITHUB_INSTALL_STATE_CACHE_PREFIX}{token}"
 
 
 def authenticated_user_id(request: Request) -> int:
@@ -92,29 +78,8 @@ def team_id_from_authorize_payload(payload: dict[str, Any] | str) -> int | None:
     return team_id_from_next_url(str(payload.get("next") or ""))
 
 
-def _flow_kind_from_legacy_payload(payload: dict[str, Any]) -> FlowKind:
-    legacy_flow = payload.get("flow")
-    if legacy_flow == "oauth_authorize":
-        return FlowKind.PERSONAL_OAUTH
-    if legacy_flow == "oauth_discover":
-        return FlowKind.OAUTH_DISCOVER
-    if legacy_flow == "team_oauth_authorize":
-        return FlowKind.TEAM_OAUTH
-    if payload.get("team_id") is not None and payload.get("next") is not None and legacy_flow is None:
-        return FlowKind.TEAM_UPDATE
-    return FlowKind.PERSONAL_INSTALL
-
-
 def authorize_state_from_payload(token: str, payload: dict[str, Any]) -> GitHubAuthorizeState:
-    flow_raw = payload.get("flow")
-    if isinstance(flow_raw, str):
-        try:
-            flow = FlowKind(flow_raw)
-        except ValueError:
-            flow = _flow_kind_from_legacy_payload(payload)
-    else:
-        flow = _flow_kind_from_legacy_payload(payload)
-
+    flow = FlowKind(payload["flow"])
     connect_from = payload.get("connect_from")
     return GitHubAuthorizeState(
         token=token,
@@ -127,13 +92,6 @@ def authorize_state_from_payload(token: str, payload: dict[str, Any]) -> GitHubA
         next_url=str(payload["next"]) if payload.get("next") else None,
         connect_from=str(connect_from) if connect_from else None,
     )
-
-
-def _legacy_team_payload(token: str, next_url: str, team_id: int | None) -> dict[str, Any]:
-    payload: dict[str, Any] = {"token": token, "next": next_url, "flow": FlowKind.TEAM_INSTALL.value}
-    if team_id is not None:
-        payload["team_id"] = team_id
-    return payload
 
 
 def store_unified_authorize_state(state: GitHubAuthorizeState, *, ttl: int | None = None) -> None:
@@ -156,31 +114,8 @@ def store_unified_authorize_state(state: GitHubAuthorizeState, *, ttl: int | Non
     cache.set(unified_authorize_pending_cache_key(state.user_id), state.token, timeout=timeout)
 
 
-def store_legacy_install_state(token: str, payload: dict[str, Any]) -> None:
-    cache.set(legacy_install_state_cache_key(token), payload, timeout=GITHUB_INSTALL_STATE_TTL_SECONDS)
-
-
 def store_personal_authorize_state(state: GitHubAuthorizeState) -> None:
-    legacy_payload: dict[str, Any] = {"user_id": state.user_id}
-    if state.connect_from is not None:
-        legacy_payload["connect_from"] = state.connect_from
-    if state.flow == FlowKind.OAUTH_DISCOVER:
-        legacy_payload["flow"] = "oauth_discover"
-    elif state.flow == FlowKind.PERSONAL_OAUTH:
-        legacy_payload["flow"] = "oauth_authorize"
-        if state.installation_id is not None:
-            legacy_payload["installation_id"] = state.installation_id
-    elif state.flow == FlowKind.TEAM_OAUTH:
-        legacy_payload["flow"] = "team_oauth_authorize"
-        if state.installation_id is not None:
-            legacy_payload["installation_id"] = state.installation_id
-        if state.team_id is not None:
-            legacy_payload["team_id"] = state.team_id
-        if state.next_url is not None:
-            legacy_payload["next"] = state.next_url
-
-    store_unified_authorize_state(state, ttl=GITHUB_INSTALL_STATE_TTL_SECONDS)
-    store_legacy_install_state(state.token, legacy_payload)
+    store_unified_authorize_state(state)
 
 
 def store_github_authorize_state(
@@ -191,12 +126,6 @@ def store_github_authorize_state(
     *,
     flow: FlowKind = FlowKind.TEAM_INSTALL,
 ) -> None:
-    legacy_payload = _legacy_team_payload(token, next_url, team_id)
-    cache.set(
-        github_authorize_state_cache_key(user_id),
-        legacy_payload,
-        timeout=GITHUB_AUTHORIZE_STATE_CACHE_TTL_SECONDS,
-    )
     store_unified_authorize_state(
         GitHubAuthorizeState(
             token=token,
@@ -210,18 +139,12 @@ def store_github_authorize_state(
 
 def peek_github_authorize_state(user_id: int, cached: Any | None = None) -> tuple[str | None, str | None, int | None]:
     if cached is None:
-        cached = cache.get(github_authorize_state_cache_key(user_id))
-    if cached is None:
         pending_token = cache.get(unified_authorize_pending_cache_key(user_id))
-        if pending_token:
-            unified = cache.get(unified_authorize_cache_key(str(pending_token)))
-            if isinstance(unified, dict):
-                return (
-                    str(unified.get("token") or pending_token),
-                    str(unified.get("next") or "") or None,
-                    team_id_from_authorize_payload(unified),
-                )
-        return None, None, None
+        if not pending_token:
+            return None, None, None
+        cached = cache.get(unified_authorize_cache_key(str(pending_token)))
+        if not isinstance(cached, dict):
+            return None, None, None
     if isinstance(cached, dict):
         token = str(cached.get("token") or "") or None
         next_url = str(cached.get("next") or "") or None
@@ -233,9 +156,7 @@ def resolve_github_setup_callback_context(
     user: User,
     state_raw: str | None,
 ) -> tuple[int | None, str | None]:
-    """Resolve redirect target from authorize cache and/or GitHub ``state`` query param."""
-    cached = cache.get(github_authorize_state_cache_key(user.id))
-    _, cached_next, cached_team_id = peek_github_authorize_state(user.id, cached=cached)
+    _, cached_next, cached_team_id = peek_github_authorize_state(user.id)
 
     team_id = cached_team_id
     next_url = cached_next
@@ -257,7 +178,6 @@ def resolve_github_setup_callback_context(
 
 
 def has_pending_personal_setup_update(user: User, installation_id: str | None) -> bool:
-    """True when the user recently started a personal GitHub configure/update flow."""
     pending_token = cache.get(unified_authorize_pending_cache_key(user.id))
     if pending_token is None:
         return False
@@ -285,20 +205,14 @@ def store_personal_manage_callback_state(user_id: int, installation_id: str) -> 
 
 
 def has_pending_team_setup_update(user: User, state_raw: str | None) -> bool:
-    """True when the user recently started a team GitHub configure/update flow."""
-    cached = cache.get(github_authorize_state_cache_key(user.id))
-    if cached is None:
-        pending_token = cache.get(unified_authorize_pending_cache_key(user.id))
-        if pending_token is None:
-            return False
-        unified = cache.get(unified_authorize_cache_key(str(pending_token)))
-        if not isinstance(unified, dict):
-            return False
-        cached = unified
+    pending_token = cache.get(unified_authorize_pending_cache_key(user.id))
+    if pending_token is None:
+        return False
+    cached = cache.get(unified_authorize_cache_key(str(pending_token)))
+    if not isinstance(cached, dict):
+        return False
 
-    expected_token, _, _ = peek_github_authorize_state(user.id, cached=cached if isinstance(cached, dict) else cached)
-    if not expected_token:
-        expected_token = str(cached.get("token") if isinstance(cached, dict) else cached)
+    expected_token = str(cached.get("token") or pending_token)
 
     if not state_raw:
         return True
@@ -311,20 +225,12 @@ def has_pending_team_setup_update(user: User, state_raw: str | None) -> bool:
 
 def load_authorize_state(token: str, *, user_id: int | None = None) -> GitHubAuthorizeState | None:
     unified = cache.get(unified_authorize_cache_key(token))
-    if isinstance(unified, dict):
-        state = authorize_state_from_payload(token, unified)
-        if user_id is not None and state.user_id != user_id:
-            return None
-        return state
-
-    legacy = cache.get(legacy_install_state_cache_key(token))
-    if isinstance(legacy, dict):
-        state = authorize_state_from_payload(token, legacy)
-        if user_id is not None and state.user_id != user_id:
-            return None
-        return state
-
-    return None
+    if not isinstance(unified, dict):
+        return None
+    state = authorize_state_from_payload(token, unified)
+    if user_id is not None and state.user_id != user_id:
+        return None
+    return state
 
 
 def consume_authorize_state(token: str, *, user_id: int | None = None) -> GitHubAuthorizeState | None:
@@ -332,7 +238,6 @@ def consume_authorize_state(token: str, *, user_id: int | None = None) -> GitHub
     if state is None:
         return None
     cache.delete(unified_authorize_cache_key(token))
-    cache.delete(legacy_install_state_cache_key(token))
     if user_id is not None:
         pending = cache.get(unified_authorize_pending_cache_key(user_id))
         if pending == token:
@@ -365,23 +270,13 @@ def consume_github_authorize_state(
     code: str | None = None,
 ) -> tuple[str, str, int | None]:
     user_id = authenticated_user_id(request)
-    cache_key = github_authorize_state_cache_key(user_id)
-    cached = cache.get(cache_key)
-    if cached is None:
-        pending_token = cache.get(unified_authorize_pending_cache_key(user_id))
-        if pending_token is not None:
-            unified = cache.get(unified_authorize_cache_key(str(pending_token)))
-            if isinstance(unified, dict):
-                cached = unified
-        if cached is None:
-            raise ValidationError("Invalid or expired state token", code="invalid_state")
+    pending_token = cache.get(unified_authorize_pending_cache_key(user_id))
+    cached = cache.get(unified_authorize_cache_key(str(pending_token))) if pending_token else None
+    if not isinstance(cached, dict):
+        raise ValidationError("Invalid or expired state token", code="invalid_state")
 
-    if isinstance(cached, dict):
-        expected_token = str(cached.get("token") or "")
-        cached_next = str(cached.get("next") or "")
-    else:
-        expected_token = str(cached)
-        cached_next = ""
+    expected_token = str(cached.get("token") or "")
+    cached_next = str(cached.get("next") or "")
 
     if not state_raw:
         if code is not None or setup_action != "update":
@@ -392,9 +287,8 @@ def consume_github_authorize_state(
         if param_token is not None and param_token != expected_token:
             raise ValidationError("Invalid or expired state token", code="invalid_state")
 
-    team_id = team_id_from_authorize_payload(cached) if isinstance(cached, dict) else None
+    team_id = team_id_from_authorize_payload(cached)
 
-    cache.delete(cache_key)
     cache.delete(unified_authorize_cache_key(expected_token))
     cache.delete(unified_authorize_pending_cache_key(user_id))
     return expected_token, param_next or cached_next, team_id
