+++ posthog/api/integration.py
@@ -23,19 +23,15 @@ from rest_framework.response import Response
 
 from posthog.api.github_callback import state as github_callback_state
 from posthog.api.github_callback.team_services import (
-    build_team_github_oauth_authorize_url,
     create_team_github_integration_from_oauth_code,
-    link_existing_team_github_integration,
     prepare_team_github_manage_callback,
 )
-from posthog.api.github_callback.types import connect_from_for_next
 from posthog.api.routing import TeamAndOrgViewSetMixin
 from posthog.api.shared import UserBasicSerializer
 from posthog.api.utils import action
 from posthog.auth import SessionAuthentication
 from posthog.domain_connect import discover_domain_connect, extract_root_domain_and_host, get_available_providers
 from posthog.exceptions_capture import capture_exception
-from posthog.models import User
 from posthog.models.instance_setting import get_instance_setting
 from posthog.models.integration import (
     ANTHROPIC_DEFAULT_INTEGRATION_ID_PREFIX,
@@ -604,8 +600,6 @@ class IntegrationViewSet(
         "patch",
         "destroy",
         "refresh_github_repos",
-        "github_link_existing",
-        "github_oauth_authorize",
         "github_prepare_callback",
     ]
     permission_classes = [TeamMemberStrictManagementPermission]
@@ -1054,18 +1048,6 @@ class IntegrationViewSet(
 
         return Response({"repositories": repositories, "has_more": has_more})
 
-    @action(methods=["POST"], detail=False, url_path="github/link_existing")
-    def github_link_existing(self, request: Request, *args: Any, **kwargs: Any) -> Response:
-        """Reuse a GitHub installation already linked to a sibling team in the same organization."""
-        instance = link_existing_team_github_integration(
-            user=cast(User, request.user),
-            organization=self.organization,
-            team_id=self.team_id,
-            source_team_id=request.data.get("source_team_id"),
-            installation_id_param=request.data.get("installation_id"),
-        )
-        return Response(self.get_serializer(instance).data)
-
     @extend_schema(request=GitHubPrepareCallbackRequestSerializer, responses={204: None})
     @action(methods=["POST"], detail=False, url_path="github/prepare_callback")
     def github_prepare_callback(self, request: Request, *args: Any, **kwargs: Any) -> Response:
@@ -1083,25 +1065,6 @@ class IntegrationViewSet(
             team_id=self.team_id,
         )
         return Response(status=204)
-
-    @action(methods=["POST"], detail=False, url_path="github/oauth_authorize")
-    def github_oauth_authorize(self, request: Request, *args: Any, **kwargs: Any) -> Response:
-        """Mint a User OAuth URL to bootstrap a fresh `code` when the install flow returns without one."""
-        installation_id = request.data.get("installation_id")
-        next_url = str(request.data.get("next") or "")
-        connect_from = (
-            request.data.get("connect_from")
-            if request.data.get("connect_from") == "posthog_code"
-            else connect_from_for_next(next_url)
-        )
-        oauth_url = build_team_github_oauth_authorize_url(
-            user_id=request.user.id,
-            team_id=self.team_id,
-            installation_id=str(installation_id),
-            next_url=next_url,
-            connect_from=str(connect_from) if connect_from else None,
-        )
-        return Response({"oauth_url": oauth_url})
 
     @extend_schema(request=None, responses={200: GitHubReposRefreshResponseSerializer})
     @action(methods=["POST"], detail=True, url_path="github_repos/refresh")
