+++ posthog/urls.py
@@ -401,11 +401,7 @@ urlpatterns = [
     # GitHub account linking (identity-only, separate from the login pipeline).
     # Must precede `social_django.urls` so the latter's `complete/<str:backend>/` doesn't swallow it.
     path("complete/github-link/", github_link_complete, name="github_link_complete"),
-    opt_slash_path(
-        "integrations/github/callback",
-        handle_setup_url,
-        name="github_team_integration_setup_callback",
-    ),
+    opt_slash_path("integrations/github/callback", handle_setup_url, name="github_team_integration_setup_callback"),
     path("", include("social_django.urls", namespace="social")),
     path("uploaded_media/<str:image_uuid>", uploaded_media.download),
     opt_slash_path("slack/interactivity-callback", posthog_code_interactivity_handler),
