+++ products/integrations/mcp/tools.yaml
@@ -103,12 +103,6 @@ tools:
     integrations-github-branches-retrieve:
         operation: integrations_github_branches_retrieve
         enabled: false
-    integrations-github-link-existing-create:
-        operation: integrations_github_link_existing_create
-        enabled: false
-    integrations-github-oauth-authorize-create:
-        operation: integrations_github_oauth_authorize_create
-        enabled: false
     integrations-github-prepare-callback-create:
         operation: integrations_github_prepare_callback_create
         enabled: false
